GPU modules fixes: - Moved kernel_modules.py to kernels\ - Updated paths to the compiled kernels. - Added init to kernels folder.
