+++ febid/kernels/__init__.py
@@ -1,0 +1,3 @@
+"""Components and kernels used by the GPU-supported FEBID simulation pipeline."""
+
+from .kernel_modules import GPU
