+++ febid/kernels/kernel_modules.py
@@ -107,17 +107,17 @@ class GPU:
         Compiled functions are stored in the class as attributes and can be called directly.
         kernels must be stored in the 'kernels' folder in the same directory as this file.
         """
-        kernels_folder = '/kernels'
-        f = open(os.path.dirname(os.path.realpath(__file__)) + os.path.join(kernels_folder, 'dep_prec_den.cl'), 'r', encoding='utf-8')
+        base_directory = os.path.dirname(os.path.realpath(__file__))
+        f = open(os.path.join(base_directory, 'dep_prec_den.cl'), 'r', encoding='utf-8')
         kernels_prec_den = ''.join(f.readlines())  # get path to kernel code
         f.close()
-        f = open(os.path.dirname(os.path.realpath(__file__)) + os.path.join(kernels_folder, 'up_surface.cl'), 'r', encoding='utf-8')
+        f = open(os.path.join(base_directory, 'up_surface.cl'), 'r', encoding='utf-8')
         kernels_up_surf = ''.join(f.readlines())  # get path to kernel code
         f.close()
-        f = open(os.path.dirname(os.path.realpath(__file__)) + os.path.join(kernels_folder, 'return_slice.cl'), 'r', encoding='utf-8')
+        f = open(os.path.join(base_directory, 'return_slice.cl'), 'r', encoding='utf-8')
         kernels_ret_slice = ''.join(f.readlines())  # get path to kernel code
         f.close()
-        f = open(os.path.dirname(os.path.realpath(__file__)) + os.path.join(kernels_folder, 'ret_slice_bool.cl'), 'r', encoding='utf-8')
+        f = open(os.path.join(base_directory, 'ret_slice_bool.cl'), 'r', encoding='utf-8')
         kernels_ret_slice_b = ''.join(f.readlines())  # get path to kernel code
         f.close()
